backend/requirements.txt
# requirements.txt
flask
google-cloud-firestore
google-cloud-storage